move vector_to_matrix
